requirements.txt
Django>=4.1,<4.2
wagtail>=5.1,<5.2
django-storages
python-dotenv==1.0.0
dj-database-url==2.1.0
uvicorn==0.23.2
pymysql
wagtail-metadata==4.0.3
wagtailmenus==3.1.9
wagtail_modeladmin==1.0.0
django-recaptcha==3.0.0